+++ src/asserts/date/index.ts
@@ -16,7 +16,7 @@ export function date(): WithBGuardType<DateSchema, Date> {
   return new DateSchema({ type: ['object'], requiredValidations: [] }) as WithBGuardType<DateSchema, Date>;
 }
 
-class DateSchema extends CommonSchema {
+export class DateSchema extends CommonSchema {
   protected _date = 1;
 
   constructor(ctx: ValidatorContext) {